requirements.txt
Django==1.9.8
Pillow==3.3.0
celery==3.1.23
coverage==4.2
djangorestframework==3.4.0
flake8<3.0.0
model_mommy==1.2.6
python-dateutil==2.5.3
pip==8.1.2
pytz==2016.6.1
requests==2.10.0
setuptools==25.1.0
wheel==0.29.0
-e .